- Difficulty slider added

+++ core/src/main/java/com/shatteredpixel/yasd/items/wands/DamageWand.java
@@ -58,8 +58,8 @@ public abstract class DamageWand extends Wand{
 	@Override
 	public String statsDesc() {
 		if (levelKnown)
-			return Messages.get(this, "stats_desc", min(), max());
+			return Messages.get(this, "stats_desc",  min(), max());
 		else
-			return Messages.get(this, "stats_desc", min(0), max(0));
+			return Messages.get(this, "stats_desc", (int) min(0), (int) max(0));
 	}
 }
